requirements.txt
alpaca==1.0.0
bokeh==3.8.1
Flask==3.1.2
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
python-dotenv==1.2.1
scikit_learn==1.7.2
ta==0.11.0
yfinance==0.2.66
